EDAS.preprocess.datatypes {EDAS}	R Documentation

EDAS Basic Preprocess - Utility Function for Setting Data Types (datetime(POSIXct)/factor/numeric/character)

Description

This utility function determines a data type for individual columns within the given data frame/data table.

Individual column of datetime(POSIXct)/factor/numeric type

.

Usage

EDAS.preprocess.datatypes(dt, factorCols = NULL, datetimeCol = NULL)

Arguments

dt
data frame or data table

factorCols
character vector defines a column list that wish to set as factor type

datetimeCol
character datetime(POSIXct) column name

Value

datetimeCol
POSIXct 으로 변환된 컬럼

numericCols
숫자로 변환된 컬럼(들)

factorCols
factor로 변환된 컬럼(들)

eval
데이터 타입이 변경된 input 데이터(data.frame/data.table)

Examples

datasets <- EDAS.getdata(c("collect_date","PowerUsage"), isModelGeneric = TRUE)
trainData <- datasets$train

colnames(trainData)
// "collect_date" "PowerUsage"   "DoW"          "isWorking"    "YI"           "Temp"         "RH"           "E"

expanded_Trainingset <- EDAS.preprocess.time.expand(trainData)
resultSet <- str(EDAS.preprocess.datatypes(expanded_Trainingset))

[Package EDAS version 1.1.1 Index]